Performing logout if the user is already connected, then calling initialize to perform an new authentication

+++ auth.js
@@ -8,19 +8,29 @@ const client = new Client({
     })
 });
 
+let hasAuthenticated = false
+
 client.on('qr', (qr) => {
     qrcode.generate(qr, { small: true });
+    hasAuthenticated = true
 });
 
 client.on('ready', () => {
-    console.log('Client is ready!');
-    client.destroy().then(() => {
-        console.log('Client destroyed successfully.');
-        process.exit();
-    }).catch((error) => {
-        console.error('Error while destroying client:', error);
-        process.exit(1);
-    });
+    if(hasAuthenticated) {
+        console.log('Client is ready!');
+        client.destroy().then(() => {
+            console.log('Client destroyed successfully.');
+            process.exit();
+        }).catch((error) => {
+            console.error('Error while destroying client:', error);
+            process.exit(1);
+        });
+    } else {
+        console.log('Performing logout');
+        client.logout().then(
+            client.initialize()
+        )
+    }
 });
 
 client.initialize();
